Add profile photo feature to user accounts
Added a new property `ProfilePhoto` to the `User` model to store the URL of the user's profile photo. Updated the `AccountController` to set a default profile photo URL (`/uploads/profile-pictures/default.png`) when a new user is created. Modified the `UserController` to include the `ProfilePhoto` property when returning user profile data. Added new endpoints `upload-profile-photo` and `delete-profile-photo` in `UserController` to allow users to upload and delete their profile photos. Updated the database schema to include the new `ProfilePhoto` column in the `UsersT` table by creating a new migration `AddProfilePhotoToUsers`. Updated the `AppDbContextModelSnapshot` to reflect the new `ProfilePhoto` column in the `UsersT` table.